--- completers/git_completer/cmd/request_pull_generated.go
@@ -1,19 +1,0 @@
-package cmd
-
-import (
-	"github.com/rsteube/carapace"
-	"github.com/spf13/cobra"
-)
-
-var request_pullCmd = &cobra.Command{
-	Use:     "request-pull",
-	Short:   "Generates a summary of pending changes",
-	Run:     func(cmd *cobra.Command, args []string) {},
-	GroupID: groups[group_interaction].ID,
-}
-
-func init() {
-	carapace.Gen(request_pullCmd).Standalone()
-	request_pullCmd.Flags().BoolS("p", "p", false, "show patch text as well")
-	rootCmd.AddCommand(request_pullCmd)
-}
